Added the ability to relocate Spectre itself.
It very much looks like there was an attempt to make this into a feature from the very beginning, with $(pwd) being used in various places instead of a hardcoded /var/spectre, as well as the whole "> Running from: $(pwd)" thing (why would this exist if $(pwd) is supposed to always be /var/spectre, anyway?) - but then I just ended up giving up halfway through and hardcoded /var/spectre everywhere else. So I can say that I finally finished this feature, ig?

This COULD technically be considered a breaking change, as some projects may rely on /var/spectre being a hard-coded path, but I chose to only increment the minor version number (ie. not indicate a breaking change, like a 1.4 release would), since:
* I have something bigger cooking up, that would make it more worthy of a 2.0 release.
* Hard-coding /var/spectre is not recommended by our docs, so it's your fault for not following them.

Now, admittedly, said docs don't (yet) exist. But hey - if they don't exist, then they can't recommend anything. So my sentence above, about something being not recommended, is technically correct. Also, the lack of docs should be a strong hint that you probably shouldn't depend on this. If, nevertheless, somebody uses this abomination of an installer and does so in such a way that the path being different from /var/spectre breaks their app, then:
* We're INCREDIBLY sorry, we really didn't expect this project to be anybody's dependency at its current stage.
* WHY??? Why would you use this? What's wrong with you? Who in their right mind would depend on some random script that doesn't even have documentation and its defaults strongly indicate that it's tied to a different project? We are INCREDIBLY flattered, of course, but... Uhhh... Are you okay?
* Their app would still be unlikely to break, as that'd also require them to have a very obscurely named envar ($SPECTRE_PATH) set, which is rather improbable.

Also, with Spectre now relocatable, it no longer needs root permissions! Yippee! (*if located in a directory that can be written to without root)

Also-also: These changes are completely untested, as we don't have any kind of test case for this code YET (magic-sauce notwithstanding, but it's a bit too bloated to be an effective test). But just like with the docs - this will change soon. With all that said, I don't THINK this has the potential to break anything? I didn't really mess around with anything that could affect global state (and state managment bugs are the most likely in such a messy (despite being 200 lines lol) codebase).